+++ src/lib.rs
@@ -2,7 +2,5 @@
 
 /// Compiler functionality.
 pub use htlk_compiler as compiler;
-/// Intermediate representation types and utilities.
-pub use htlk_ir as ir;
 /// Runtime functionality.
 pub use htlk_rt as rt;